requirements-2.txt
playsound3
soundcloud-lib==0.6.1
PyPI
sclib
pypi
SClib
opencv-python
absl-py==2.3.1
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1

datasets==4.4.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
evaluate==0.4.6
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.123.10
fastjsonschema==2.21.2
ffmpy==1.0.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.61.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.6.0
google-pasta==0.2.0
gradio==6.0.2
gradio_client==2.0.1
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.5
idna==3.11
ipykernel==7.1.0
ipython==8.37.0
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keras==3.12.0
kiwisolver==1.4.9
lark==1.3.1
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.45.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.5.0
notebook_shim==0.2.4
numba==0.62.1
numpy==1.26.4
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==4.25.8
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.29.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
tomli==2.3.0
tomlkit==0.13.3
torch==2.2.2
torchaudio==2.2.2
torchcodec==0.0.0.dev0
torchvision==0.17.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers.git@06d52fe106dc6922815dc25bfa156a9d78b16959
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
widgetsnbextension==4.0.15
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2025.11.12

